+++ main.tf
@@ -33,7 +33,8 @@ module "db_parameter_group" {
   description     = var.parameter_group_description
   family          = var.family
 
-  parameters = var.parameters
+  parameters   = var.parameters
+  skip_destroy = var.parameter_group_skip_destroy
 
   tags = merge(var.tags, var.db_parameter_group_tags)
 }
@@ -49,7 +50,8 @@ module "db_option_group" {
   engine_name              = var.engine
   major_engine_version     = var.major_engine_version
 
-  options = var.options
+  options      = var.options
+  skip_destroy = var.option_group_skip_destroy
 
   timeouts = var.option_group_timeouts
 
@@ -63,25 +65,36 @@ module "db_instance" {
   identifier            = var.identifier
   use_identifier_prefix = var.instance_use_identifier_prefix
 
-  engine            = var.engine
-  engine_version    = var.engine_version
-  instance_class    = var.instance_class
-  allocated_storage = var.allocated_storage
-  storage_type      = var.storage_type
-  storage_encrypted = var.storage_encrypted
-  kms_key_id        = var.kms_key_id
-  license_model     = var.license_model
+  engine                   = var.engine
+  engine_version           = var.engine_version
+  engine_lifecycle_support = var.engine_lifecycle_support
+  instance_class           = var.instance_class
+  allocated_storage        = var.allocated_storage
+  storage_type             = var.storage_type
+  storage_encrypted        = var.storage_encrypted
+  kms_key_id               = var.kms_key_id
+  license_model            = var.license_model
 
   db_name                             = var.db_name
   username                            = var.username
   password                            = var.manage_master_user_password ? null : var.password
   port                                = var.port
   domain                              = var.domain
+  domain_auth_secret_arn              = var.domain_auth_secret_arn
+  domain_dns_ips                      = var.domain_dns_ips
+  domain_fqdn                         = var.domain_fqdn
   domain_iam_role_name                = var.domain_iam_role_name
+  domain_ou                           = var.domain_ou
   iam_database_authentication_enabled = var.iam_database_authentication_enabled
   custom_iam_instance_profile         = var.custom_iam_instance_profile
   manage_master_user_password         = var.manage_master_user_password
   master_user_secret_kms_key_id       = var.master_user_secret_kms_key_id
+
+  manage_master_user_password_rotation                   = var.manage_master_user_password_rotation
+  master_user_password_rotate_immediately                = var.master_user_password_rotate_immediately
+  master_user_password_rotation_automatically_after_days = var.master_user_password_rotation_automatically_after_days
+  master_user_password_rotation_duration                 = var.master_user_password_rotation_duration
+  master_user_password_rotation_schedule_expression      = var.master_user_password_rotation_schedule_expression
 
   vpc_security_group_ids = var.vpc_security_group_ids
   db_subnet_group_name   = local.db_subnet_group_name
@@ -89,12 +102,14 @@ module "db_instance" {
   option_group_name      = var.engine != "postgres" ? local.option_group : null
   network_type           = var.network_type
 
-  availability_zone   = var.availability_zone
-  multi_az            = var.multi_az
-  iops                = var.iops
-  storage_throughput  = var.storage_throughput
-  publicly_accessible = var.publicly_accessible
-  ca_cert_identifier  = var.ca_cert_identifier
+  availability_zone      = var.availability_zone
+  multi_az               = var.multi_az
+  iops                   = var.iops
+  storage_throughput     = var.storage_throughput
+  publicly_accessible    = var.publicly_accessible
+  ca_cert_identifier     = var.ca_cert_identifier
+  dedicated_log_volume   = var.dedicated_log_volume
+  upgrade_storage_config = var.upgrade_storage_config
 
   allow_major_version_upgrade = var.allow_major_version_upgrade
   auto_minor_version_upgrade  = var.auto_minor_version_upgrade
@@ -132,6 +147,9 @@ module "db_instance" {
   create_cloudwatch_log_group            = var.create_cloudwatch_log_group
   cloudwatch_log_group_retention_in_days = var.cloudwatch_log_group_retention_in_days
   cloudwatch_log_group_kms_key_id        = var.cloudwatch_log_group_kms_key_id
+  cloudwatch_log_group_skip_destroy      = var.cloudwatch_log_group_skip_destroy
+  cloudwatch_log_group_class             = var.cloudwatch_log_group_class
+  cloudwatch_log_group_tags              = var.cloudwatch_log_group_tags
 
   timeouts = var.timeouts
 
@@ -141,7 +159,8 @@ module "db_instance" {
   restore_to_point_in_time = var.restore_to_point_in_time
   s3_import                = var.s3_import
 
-  tags = merge(var.tags, var.db_instance_tags)
+  db_instance_tags = var.db_instance_tags
+  tags             = var.tags
 }
 
 module "db_instance_role_association" {